Moved /is admin setbanklimit command to the bank module
