+++ src/main/java/com/bgsoftware/superiorskyblock/module/bank/BankModule.java
@@ -5,12 +5,7 @@ import com.bgsoftware.superiorskyblock.SuperiorSkyblockPlugin;
 import com.bgsoftware.superiorskyblock.api.commands.SuperiorCommand;
 import com.bgsoftware.superiorskyblock.module.BuiltinModule;
 import com.bgsoftware.superiorskyblock.module.IModuleConfiguration;
-import com.bgsoftware.superiorskyblock.module.bank.commands.CmdAdminDeposit;
-import com.bgsoftware.superiorskyblock.module.bank.commands.CmdAdminWithdraw;
-import com.bgsoftware.superiorskyblock.module.bank.commands.CmdBalance;
-import com.bgsoftware.superiorskyblock.module.bank.commands.CmdBank;
-import com.bgsoftware.superiorskyblock.module.bank.commands.CmdDeposit;
-import com.bgsoftware.superiorskyblock.module.bank.commands.CmdWithdraw;
+import com.bgsoftware.superiorskyblock.module.bank.commands.*;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.event.Listener;
 
@@ -77,7 +72,7 @@ public class BankModule extends BuiltinModule<BankModule.Configuration> {
 
     @Override
     public SuperiorCommand[] getSuperiorAdminCommands(SuperiorSkyblockPlugin plugin) {
-        return new SuperiorCommand[]{new CmdAdminDeposit(), new CmdAdminWithdraw()};
+        return new SuperiorCommand[]{new CmdAdminDeposit(), new CmdAdminSetBankLimit(), new CmdAdminWithdraw()};
     }
 
     @Override
